# Optimize alignment dynamic programming by small code restructuring