Reduce the number of branches in the most inner loop.

Before:
```
$ time ./target/release/de-novo-align --peptides rustyms/data/200305_HER_test_04_DENOVO_excerpt.csv --database examples/de-novo-align/database.fasta --out-path /tmp/eta.csv --no-parallel

real	0m3,969s
user	0m3,949s
sys	0m0,012s
```

After:
```
$ time ./target/release/de-novo-align --peptides rustyms/data/200305_HER_test_04_DENOVO_excerpt.csv --database examples/de-novo-align/database.fasta --out-path /tmp/eta.csv --no-parallel

real	0m2,913s
user	0m2,892s
sys	0m0,012s
```